# br-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Quote scheduling and notification system",
        "title": "Quote Scheduler API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/jobs": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Assign a quote to a technician",
                "parameters": [
                    {
                        "description": "Assignment",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.assignJobRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.Response-service_JobResponse"
                        }
                    },
                    "409": {
                        "description": "Technician has a conflicting job",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/jobs/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get a job by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response-service_JobResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/jobs/{id}/complete": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Mark a job as complete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Technician",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.completeJobRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response-service_JobResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/notifications/stream": {
            "get": {
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Stream notifications via SSE",
                "parameters": [
                    {
                        "type": "string",
                        "description": "technician or manager",
                        "name": "recipient_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Recipient UUID",
                        "name": "recipient_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/notifications/{id}/read": {
            "patch": {
                "tags": [
                    "notifications"
                ],
                "summary": "Mark a notification as read",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notification ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Recipient UUID",
                        "name": "recipient_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/quotes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quotes"
                ],
                "summary": "List unscheduled quotes",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response-array_models_Quote"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quotes"
                ],
                "summary": "Create a quote",
                "parameters": [
                    {
                        "description": "Quote",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Quote"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.Response-models_Quote"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/technicians": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "technicians"
                ],
                "summary": "List technicians",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response-array_models_Technician"
                        }
                    }
                }
            }
        },
        "/technicians/{id}/jobs": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "technicians"
                ],
                "summary": "Get a technician's job schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Technician ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date (YYYY-MM-DD), defaults to today",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response-array_service_JobSlotResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "handler.Response-array_models_Quote": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Quote"
                    }
                }
            }
        },
        "handler.Response-array_models_Technician": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Technician"
                    }
                }
            }
        },
        "handler.Response-array_service_JobSlotResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.JobSlotResponse"
                    }
                }
            }
        },
        "handler.Response-models_Quote": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Quote"
                }
            }
        },
        "handler.Response-service_JobResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/service.JobResponse"
                }
            }
        },
        "handler.assignJobRequest": {
            "type": "object",
            "required": [
                "manager_id",
                "quote_id",
                "starts_at",
                "technician_id"
            ],
            "properties": {
                "manager_id": {
                    "type": "string"
                },
                "quote_id": {
                    "type": "string"
                },
                "starts_at": {
                    "type": "string"
                },
                "technician_id": {
                    "type": "string"
                }
            }
        },
        "handler.completeJobRequest": {
            "type": "object",
            "required": [
                "technician_id"
            ],
            "properties": {
                "technician_id": {
                    "type": "string"
                }
            }
        },
        "models.JobStatus": {
            "type": "string",
            "enum": [
                "scheduled",
                "completed",
                "cancelled"
            ],
            "x-enum-varnames": [
                "JobStatusScheduled",
                "JobStatusCompleted",
                "JobStatusCancelled"
            ]
        },
        "models.Quote": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "customerName": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.QuoteStatus"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.QuoteStatus": {
            "type": "string",
            "enum": [
                "unscheduled",
                "scheduled"
            ],
            "x-enum-varnames": [
                "QuoteStatusUnscheduled",
                "QuoteStatusScheduled"
            ]
        },
        "models.Technician": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "service.JobResponse": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "ends_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "manager_id": {
                    "type": "string"
                },
                "quote_id": {
                    "type": "string"
                },
                "starts_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.JobStatus"
                },
                "technician_id": {
                    "type": "string"
                }
            }
        },
        "service.JobSlotResponse": {
            "type": "object",
            "properties": {
                "ends_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "starts_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.JobStatus"
                }
            }
        }
    }
}